fix memory calculation for 32bit

+++ test/runtests.jl
@@ -121,7 +121,7 @@ end
 # determine parallelism
 if !set_jobs
     jobs = Sys.CPU_THREADS
-    memory_jobs = Int(Sys.free_memory()) ÷ (2 * 2^30)
+    memory_jobs = Int64(Sys.free_memory()) ÷ (2 * 2^30)
     jobs = max(1, min(jobs, memory_jobs))
 end
 @info "Running $jobs tests in parallel. If this is too many, specify the `--jobs` argument to the tests, or set the `JULIA_CPU_THREADS` environment variable."
